Split dapparatus off into its own component
made App a functional component
